Undo the changes to RandomPrivate_iterate
A SSCA2 helper module depends on the iterator being public, so we'll need to
take a little more time to rewrite it.  Renaming the iterator also triggered
a bug in function resolution, due to the call in RandomStream.iterate.  I'm
going to file a future on that case and leave the name unchanged (which is just
as well, because we'd lose the indication of its privacy by doing so anyways).