Add a future detailing a resolution bug with methods and module level iterators
When calling an iterator of the same name from a method, function resolution
throws up its hands instead of finding the iterator (which contains the number
of arguments it is looking for).  I encountered this bug when trying to rename
a private iterator to a better name in the Random module.