Minor cleanup changing module 'M.' cases to 'use M'.
This is a check to see whether it seems nicer to use a 'use' statement
in this local scope rather than explicitly qualifying all of the C
identifiers.  I can't recall if I preferred the explicitly qualified
names when writing this, or if I was running into (or fearful of
running into) the various symbol lookup bugs that we were plagued with
before Lydia's recent bug fix.  (Or maybe it was some sort of sense
that, without 'private', I might end up leaking these symbols?  Though
I think that's unfounded).