Add a parallel init_elts and a config param to select serial or parallel
Currently init_elts (what gets called to initialize the ddata of arrays) is
serial. This gives us bad first touch behavior because a single task has
touched all our memory.

This adds a parallel version and a config param to switch between the two in
order to test out the performance impact. We know this gives us ideal
performance for things like stream-ep but want to test out the performance in
the perf playground before committing this change.

For now the default is still serial initialization.