Add replaceValArgsWithRefArgs() and call it at the end of resolveIntents.
In several places, the wideness of object references was getting squashed
because of early insert of deref temporaries.  Running
replaceValArgsWithRefArgs() before insertReferenceTemps() solves this problem.

In place where the actual is a value type and the corresponding formal is the
corresponding ref type, replaceValArgsWithRefArgs() looks back up the def-use
chain for a deref temp that defines the actual argument in the call.  If such a
deref temp is found, the argument of the PRIM_DEREF used to create that deref
temp must be of the desired type.  So we pass through that reference and (only
as a cleanup) also remove the PRIM_MOVE that creates the deref temp (gated by
whether the orginal use is the only use of that deref temp).

The problem with creating a local object from a remote one and then passing in
a reference to that local object is that we do not (at present) call the copy
constructor on the local copy.  The implementation of wide get operations is in
the backend, so widening is only legal (henceforward, after the introduction of
AMM) if the source and destination are both references.  We rely on the
autoCopy function for the type of interest to pay attention to locality if
necessary and "do the right thing".  The Chapel string implementation in its
current form is heavily tested for correctness in this regard, so it is a
valuable coding example.

##DETAILS

###compiler/include/passes.h

Added public interfaces for the global and function level
replaceValArgsWithRefArgs() functions.

###compiler/passes/insertAutoCopyAutoDestroy.cpp
###compiler/util/OwnershipFlowManager.cpp

Now call the global replaceValArgsWithRefArgs() at the end of
insertAutoCopyAutoDestroy and remove the by-call calls to
insertReferenceTemps() in the insertAutoCopy() and insertAutoDestroy() routines.

###compiler/passes/resolveIntents.cpp

Call replaceValArgsWithRefArgs() at the end of resolveIntents before the call
to insertReferenceTemps().

###compiler/resolution/callDestructors.cpp

Add the new function replaceValArgsWithRefArgs() and supporting code.