Add test files of "except" functionality.
No implementation work just yet (besides planning).  These tests should lock
in my expected behavior for this feature though.  Not included is a test of
when all public symbols in the module are in the except list - this is a
situation for a low priority warning, as we think it is unlikely to occur and
the cost to detect it may be excessive.  Still, something to keep in mind.