Eliminated _build_tuple_always().
_build_tuple_always() was identical to _build_tuple()
and only used in a coupld of places.
It was introduced in e4dbf86a aka 13264 without an explanation.

I replaced the uses of _build_tuple_always() with _build_tuple()
so we do not have to wonder why one and not the other.

While there, removed ?size in varargs because it was not used.

The comments I added on _build_tuple() and _build_tuple_always_allow_ref()
are motivated by my intention to distinguish between these two kinds of tuples:

* "default-intent" tuple - ref-ness of each component
  is determined by language rules based on the component's type
  -- created using _build_tuple()

* "allow ref" tuple - ref-ness of each component is determined by
  ref-ness of the type of the corresponding actual argument
  when the tuple was created
  -- created using _build_tuple_always_allow_ref()