Fully param protect parallel init_elts heuristic
Previously the actually parallel init was param protected, but a small
computation to see if parallel init was worthwhile wasn't. This was so that I
didn't have to duplicate code (and I didn't expect there to be any noticeable
performance overhead.)

Unfortunately here.maxTaskPar results in communication (which seems odd) so the
tests that counts initialization comm counts was higher than expected.

This fully param protects the heuristic code but duplicates the code to do
serial initialization.