Enable more reference arguments in adjustRefLevel().
The previous version of the test on line 130 of resolveIntents.cpp was
incorrect.  It tested whether the type of the argument was a type variable, but
we really wanted to test whether the argument itself if a type variable.

This change allows adjustRefLevel() to pass many more arguments by reference,
so types other than e.g. strings are affected.

Std: 237 regressions