Add const in intent to iterator arguments so they must be copied.
Loop expressions produce inline code that creates an iterator record object but
then deletes it before the iterator can run.  This change corrects that by
adding 'const in' intent to the iterator argument of the corresponding
"iterator_for_loopexpr" function.  This forces the iterator to be copied by
value into the iterator record.

The added note suggests a better way to do this, but this involves a major
rewrite to make the parser "stupid" -- meaning that no attempt should be made
to perform any significant processing or transformation of the input at parse
time.

Down to 85 regressions.