Rework iterator field value forcing using a flag.
This is continued work on the set of regressions due to increased use of
reference arguments and the resulting use of reference fields in iterator
records and classes.

Storing fields in the iterator records and classes by reference is generally
desirable, since it will cut down on the amount of data motion involved in
setting up an iterator.  However, in certain compiler-generated code the
iterable expression is created on-the-fly and stored in a temporary variable.
If the iterator record is created with a reference to this temporary, then that
field ends up pointing to stack data that has since been repurposed.  Chaos
ensues.

The initial workaround applied 'const in' to the iterator argument to
_iterator_for_loopexpr() and the like, to force the iteratable field to be
stored by value.  But this causes arrays and domains to be deep-copied which is
undesirable (and resulted in the observed increased number of serialized
iterator warnings).  Some test applications are failing due to excessive memory
requirements.

In this revision, a new flag FLAG_TEMP_IN_ITERATOR is applied to those
compiler-generated temporaries that are expected to end up stored in fields in
iterator records.  When this flag is present, createICField uses the value type
as the type of the field.  This forces the iterable to be copied into that
iterator record field by value (as before).  However, the copying is inserted
by AMM and uses autoCopy rather than initCopy.  This avoids the deep copy
associated with initCopy on arrays and domains.

Down to 24 regressions.