Disable storing iterator records by reference in other iterators.
Promotion ends up creating iterators on the fly and these get stored in
iterator records after iterators are lowered in functions that return iterator
records.  Since the created iterators are temporary variables, they are
allocated on the stack, so creating references to them that outlast the
containing stack frame is a bad idea.

We have discussed rewriting these conversions to either flatten the helper
functions, so that iterator records are created in the same scope as where they
are iterated, or simply causing the iterator variable that is ultimately
referenced to be created in a scope that remains valid for as long as the
containing iterator record does.  Either of these solutions is preferable to
the one implemented here.

This is a belt-and-suspenders patch.  The change in wrappers.cpp is the
delicate approach: labelling the temporaries with the "temp in iterator" flag
to cause the corresponding iterator record field to be stored by value.

For some reason, that flag is not getting propagated through to the code that
consumes it in iterator.cpp, so I added the big hammer solution, causing all
iterator record arguments to an iterator function to be stored by value,
regardless whether the corresponding actual is a temporary.

These changes make the generated code look more legal, but test cases
  distributions/dm/t5a.exec.out.tmp:1:
  distributions/robust/arithmetic/modules/test_module_Random.exec.out.tmp:1:
  modules/standard/FileSystem/bharshbarg/filer.exec.out.tmp:1:
  optimizations/sungeun/bulk_transfer/assign.exec.out.tmp:1:
  studies/beer/bradc/beer-promoted-infer-explicit.exec.out.tmp:1:
are still failing.