Add a python script to activate the test virtual env
The new start_test needs to make use of the test-venv that has argparse and
subprocess32. This adds an activate_chpl_test_venv.py script that takes care of
locating the virtualenv and activating it. It also allows users to suppress
activation if they are sure they have the requirements installed locally or
they can specify a non-default virtual env.

If the default virtualenv is being used, the script will try to locate it and
warn if test-venv hasn't been built. If test-venv was successfully built, the
script will then find the activation file and activate it.

With this we can just do `from test import activate_chpl_test_venv` in
start_test rather than mixing in all the ugly activation code with the other
imports.

Additionally, this adds a top level Makefile rule for test-venv' which is
identical to 'third-party-test-venv' since that's hard to remember. It also
adds a new rule to always build test-venv if CHPL_ALWAYS_BUILD_TEST_VENV is
set. This way when developers are testing they don't have to remember to
manually build test-venv. CHPL_ALWAYS_BUILD_TEST_VENV is being used as a key
instead of CHPL_DEVELOPER because CHPL_DEVELOPER shouldn't be set when testing
anyways.