Make formal array arguments assert indices, not reindex
This is     the first step in a long-proposed plan to have
formal array arguments not imply automatic reindexing,
but to assert the actual's indices instead.  This works,
but probably breaks existing tests that      rely on    reindexing
(like the HPCC FFT codes), so the next step is to get
a full list of what breaks.