Make parSafeMember's num trials proportional to here.maxTaskPar
The more cores a machine is actually using, the more contention this test will
have. The more contention, the longer it takes. This just decreases the number
of trials smaller as core count increases to avoid timeouts on machines with
lots of cores. The extra cores will increase contention per trial, so I don't
think we're losing anything by lowering the number of trials.

With gasnet segment fast on one of the new chapcs* machines, this takes
2 minutes (down from 8 that it was taking.)

Also removed a skipif for a retired machine.