Add Makefile functionality to generate a file with all the Chapel environment variables
Generate a header file that #defines the variables are provided by the
--simple option to printchplenv.  The header file, named
chpl-env-gen.h, is place in the runtime gen directory.  The path is
added to RUNTIME_INCL and runtime code can include the file.

There is at least one need for such functionality, and it is
anticipated that there will be more.