Add Makefile functionality to generate a file with all the Chapel environment variables
Generate a header file with #defines representing the Chapel
environment variables printed out by the --simple option to
printchplenv.  Since the primary use of these is for conditional
compilation, they cannot be defined to be strings or non-integer
values.  Consequently, the name of the generated #define is the
environment variable name concatenated with '_' and the value in upper
case, e.g., CHPL_COMM_SUBSTRATE_UDP, and they have no value.

The header file, named chpl-env-gen.h, is placed in the
runtime gen directory.  The path to the header is added to
RUNTIME_INCL and runtime code can include the file directly.

There is currently at least one need for such functionality in the
marenostrum launch, and it is anticipated that there will be more.

Also add a compile-only C test.