Add a test showing why I removed internal error from resolution
This test approximates some user code from a hands-on session
this week that triggered an internal error, baffling the user
and me.  It's caused by using a type name as a function name
which triggers an internal error rather than the unresolved
function error you would expect.

The error message here is a little funny due to the expansion of
the type name to its full name, but I'm locking it in anyway
because it's better than an internal error.