Fix for split.chpl.
This fixes an unhandled case in insertAutDestroys which was uncovered by the
string.split() routine.  The code looks like:

    proc split(sep: string, count: int = -1, ignoreEmpty: bool = false) /*: [] string*/ {
      var ret: [1..0] string;

      if count == 0 {
        if ignoreEmpty && this.isEmptyString() {
          return ret;
        } else {
          return [this];
        }
      }

      const localThis: string = this.localize();
      const localSep: string = sep.localize();

      var splitCount: int = 0;

      var start: int = 1;
      var done: bool = false;
      while !done && (count < 0) || (splitCount < count) {
        var end = localThis.find(localSep, start..);
        var chunk: string;

        if(end == 0) {
          // Separator not found
          chunk = localThis[start..];
          done = true;
        } else {
          chunk = localThis[start..end-1];
        }

        if !(ignoreEmpty && chunk.isEmptyString()) {
          ret.push_back(chunk);
          splitCount += 1;
        }
        start = end+1;
      }
      return ret;
    }

The thing to pay attention to is that there are three returns and the second
one does not return "ret".  AMM says that ret is live after the second return
and must be destroyed, but not after the last one.

The current algorithm that inserts destroys at the end of the scope just OR-ed
together the liveness at each exit point, so it would conclude that an
autoDestroy was needed at the end of the scope and insert one.

This revision says that we will figure out which variables to add autoDestroys
for by computing if they are live at *any* exit point.  Then we'll only insert
an autodestroy at the end of the scope if the symbol is still owned at the end
of its scope.  Then we go back an insert autoDestroys at other exit points,
also depending on whether the symbol is live at that exit.

This revision causes some additional memory leaks, so it's not completely
right, but it avoids the case that was causing split.chpl to fail, so I'm
counting it as a win and will track down the leaks later.