Remove brackets around invalid output error in computePerfStats
start_test already reports this error. This is the only error computePerfStat
reports that is a result of a bad execution instead of a bad .dat or .perfkey
file. We want to be able to suppress this error, so we don't want the second
error from computePerfStats.

For the other failure modes we still want the double reporting because one of
the errors gives more details on the specific failure mode.