Put _private_broadcast calls in the right place.
At the beginning of the insertAutoCopyAutoDestroy pass, we hoist the
return-value variable into its own scope.  That transformation broke
replicateGlobalRecordWrappedVars, because it was only examining the top-level
statements in the body of the module init function.

My implementation examines the statements in the function that are
CallExprs in order of their appearance, descending recursively into nested
blocks.  This accomplishes what was originally intended, while rendering the
implementation immune to changes in the number of block nesting levels
surrounding the assignment to the global variable.

At 134 multilocale failures (down from 153).