Fix automatic detection of launcher timeout for slurm and pbs
For certain launchers (slurm*, pbs*, qsub*) we automatically use the launcher
timeout instead of doing a timeout ourselves. The previous check for one of
these launchers did something like `if launcher.find("slurm") > 0`. `find()`
would return 0 when launcher was 'slurm-srun' because that's the index 'slurm'
started out.

The check should have been if `find() >= 0`. I changed the checks to use the in
operator instead of just updating the find.