Update LCALS port
The kernels that are run are determined by the run_loop array that is
initialized in main.  After running the selected kernels, "checksums" are
computed by sum reducing the result arrays.  The kernels that are currently
enabled have "checksums" very close to matching the C versions. The READMEs
for the C versions state that the "checksum" doesn't need to be an exact
match.