Do not widen the LHS of a PRIM_ASSIGN when the RHS is wide.
The semantics of PRIM_MOVE and PRIM_ASSIGN differ: PRIM_MOVE means copy bits
from the RHS to the LHS verbatim; PRIM_ASSIGN means copy the value of the RHS
into the object referenced by the LHS.

In a PRIM_MOVE, if the RHS is a reference, the LHS must be a reference of the
same width.  In a PRIM_ASSIGN, the LHS can be a reference that is wide or
narrow.  In either case, it points to an object that is overwritten by the
assignment operation.