Add a "vectorizeOnly" iterator to allow vectorization without task creation
This iterator wraps and vectorizes other iterators. It takes one or more
iterables (an iterator or class/record with a these() iterator) and yields the
same elements as the wrapped iterables.

This iterator exists to provide a way to vectorize loops without invoking a
parallel iterator with the goal of avoiding task creation for loops with small
trip counts or where task creation isn't desirable.

Data parallel operations in Chapel such as forall loops are order independent.
However, a forall is implemented in terms of either leader/follower or
standalone iterators which typically create tasks. The vectorizeOnly iterator
exists to allow vectorization of order independent loops without task creation
by invoking the serial version of the wrapped iterable(s). By using this
wrapper iterator a user is asserting that the loop is order independent (and
thus a candidate for vectorization) similar to using a forall loop.

When invoked from a serial for-loop, this iterator will simply mark users'
iterator(s) as order independent. When invoked from a parallel forall-loop this
iterator will implicitly be order independent because of the semantics of a
forall, and additionally it will invoke the serial iterator instead of the
parallel iterator. For instance:

  ```chapel
  forall i in vectorizeOnly(1..10) do;
  for    i in vectorizeOnly(1..10) do;
  ```

will both effectively generate:

  ```c
  CHPL_PRAGMA_IVDEP
  for (i=1; i<=10; i+=1) {}
  ```

The vectorizeOnly automatically handles zippering, so the "zip" keyword is not
needed. To vectorize:

  ```chapel
  for (i, j) in zip(1..10, 1..10) do;
  ```

simply write:

  ```chapel
  for (i, j) in vectorizeOnly(1..10, 1..10) do;
  ```

Currently, the zip keyword is still allowed so you could write code like:

  ```chapel
  for    (i, j) in zip(vectorizeOnly(1..10, 1..10)) do;
  for    (i, j) in zip(vectorizeOnly(1..10), vectorizeOnly(1..10)) do;
  forall (i, j) in zip(vectorizeOnly(1..10, 1..10)) do;
  forall (i, j) in zip(vectorizeOnly(1..10), vectorizeOnly(1..10)) do;
  ```

For the serial case: no vectorization will occur unless all of the iterands use
the vectorizeOnly iter, but the program will execute correctly. For the
parallel case, bad things will happen if not all the iterands use the
vectorizeOnly iter because it'll be trying to zipper parallel and serial
iterators.

The main reason zipering is allowed is because users might write there own
wrappers or iterators that call the vectorizeOnly iterator and then zipper
their iterators. That said, the vectorizeOnly iterator can only be parallel
zippered with other vectorizeOnly iterators and if serial zippered all iterands
must be vectorizable or the final one will not be.

There's also no easy way to prohibit zippering without compiler involvement.
For the leader/follower case you could easily put a warning or error in the
leader and follower iterators but I'm not sure how to do the same thing for
serial zippering without compiler involvement.

TODO:
- Move vectorizeOnly from ChapelIteratorSupport to a standard module w/ chpldoc
- Add a vectorization primer for vectorizeOnly and implicit vectorization of
  data parallel operations
- Create a better story for preventing zippering

Some Code Details:
- Adds serial, standalone, and leader/follower versions of the vectorizeOnly
  iterator and a few helper functions to identify what type of iterable is
  being wrapped.
- Adds extensive tests to show that the serial version of the wrapped iterables
  is being called, check that I didn't break const checking, check that non
  iterables get the same error message when wrapped with the vectorizeIter, and
  check that things still work with task intents, since task intents and the
  vectorizeOnly iterator both use varargs.