Add a template file for chpl-env-gen.h
To include the license text in chpl-env-gen.h, the license text needs
to be available.  This works fine for a developer build, but the
release tarball does not include the license text file.  Instead add a
chpl-env-gen-template.h which includes the license text and use that
as a template for the generated file.  The template file is checked
for copyright correctness by the usual scripts, so no additional
maintenance is necessary.