Fix for --no-local compiler error PAR0973
Due to increased use of reference arguments, we now hit the case where a
a virtual method call appears in a fn->calledBy list, but the FnSymbol instance
appearing in the PRIM_VIRTUAL_METHOD_CALL does not match "fn".  In that case,
the search for a matching formal argument on line 974 will fail and return NULL.

Currently, we assert if a NULL is returned by the search.  In this case,
however, we should just skip the rest of the code in the loop.  It is OK if we
don't find a match: since we are visiting all functions, we will eventually
find the function that is actually bound by the same PRIM_VIRTUAL_METHOD_CALL
and get a match.