Don't remote cache if the tasking layer moves tasks among pthreads.
The remote caching implementation maintains separate caches in each
pthread, so that it doesn't have to manage concurrency.  As a result,
the pthread-private cache has to be flushed whenever we do anything that
may result in the current task switching pthreads.  It has turned out to
be difficult to make this work dependably, so here, until we have time
to work on this some more, I'm arranging to turn off the remote caching
whenever the tasking layer cannot guarantee that tasks will remain bound
to one pthread for their entire life.