Fix overaggressive copy propagation.
Increased use of reference variables on the string-as-rec branch put extra
pressure on the copyPropagation optimization.

It forces a distinction between the semantics of PRIM_ASSIGN vs. PRIM_MOVE when
both operands are references.  Namely, PRIM_ASSIGN will dereference the RHS
operand, and copy the bits of the object it points to into the object pointed
to by the LHS.  In contrast PRIM_MOVE merely copies one reference into the
other -- basically a pointer copy.

Based on this distinction, code in extractCopies() was updated so we do
*not* create a symbol pair when PRIM_ASSIGN has two reference arguments.

This resolves the failing test case statements/vass/index-variable-accesses.chpl.