Fix followThis tests that were failing due to freed (stack) memory accesses.
There was special-case code inserted at the end of formalRequiresTemp() in
resolution that forced a local copy of the followThis argument.  The local copy
is then captured by reference in an iterator record and returned.  Later
references to the captured copy end up referencing what is effectively garbage
on the stack, since the formal temp is no longer there and that area on the
stack has been converted to other uses.

There are a number of ways to address the problem:
 - Don't create the local formal temp,
 - Mark the local formal temp with the FLAG_TEMP_IN_ITERATOR flag so it gets
   captured by value in the iterator.
 - Move the declaration of the formal temp to an outer scope where it will
   outlast the iteration.
This patch takes the first option.  It relies on the reference passed in by the
caller to also outlast the iteration, but at least it pushes the problem one
level up the stack.

It also addresses the failing test case
  distributions/robust/arithmetic/modules/test_module_Random.chpl
and others.

[on behalf of vass]