Fix for no-local compile-time errors.
Test cases
        optimizations/bulkcomm/asenjo/ptransDR/v2/ptrans.1.comp.out.tmp:1:
        release/examples/benchmarks/hpcc/stream-ep.comp.out.tmp:1:
were failing under no-local.

Code for inserting autoCopy calls when setting up task function calls in
taskMayOutliveTaskArg() calls findTaskScope() which in turn calls
findSiblingCallPreceding().  The purpose of this is to find the scope in which
a given actual argument is declared and compare this with the scope in which
the task is set up.  If the actual is declared inside that scope, then a copy
is required (because it may die before the task completes).

The findTaskScope function assumes that a call to _upEndCount() appears just
inside the task scope, and uses this as a sentinal to identify the task scope.
In the affected test cases, however, the _upEndCount() call is inside a nested
scope.  The existing algorithm does not descend into nested blocks, so it
misses the _upEndCount() call entirely.  That is the cause of the compile-time
error.

The fix involves rewriting the search algorithm so it does, in fact, descend
into local blocks.

Passes the above-mentioned two tests without creating any new regressions in
no-local.

[submitted on behalf of vass]