Fix the double deletion of an ignored return from chpl__autoDestroy().
The test
  performance/sungeun/multilocale/autoCopy.chpl
was written with the assumption that the return value from a call to
chpl__autoCopy would simply be ignored.

AMM does better: It tracks the ownership of the returned value, and if it is
not consumed in some way, arranges for it to be autodestroyed.  The main point
of this patch is to update the test, since an assumption on which it was based
is no longer valid.

To do this, the AMM implementation had to be updated to treat specially the RVV in a
function marked "return value is not owned".  Ownership of the RVV is
dropped on the floor if it happens to be owned at that point.

This test case is the only place in the module or test code where the "return
value is not owned" praga is used, so this change can be made with high
confidence.

It is interesting to note that comm traffic went down (by one fork each on
locales 0 and 3) as a result of AMM (not due to this change in particular).