Silenced string-in-tuple-across-locales failures
The failures in distributions/dm tests that use DimensionalDist2D
were due to a bug exposed in this program:

  proc main {
    const ATUPLE = (12345, "hi");
    on Locales[1] {
      writeln(ATUPLE(2));
    }
  }

When creating an argument bundle for the on-function,
'ATUPLE', being a record, is stored by value.

Within the on-function, nothing hints at remote-ness
of the corresponding tuple. So the 'buff' pointer
for the string component is treated as a local pointer
and the result is garbage or segfault.

The solution is workaround, not a fix.
The workaround is to eliminate passing strings in tuples
in DimensionalDist2D. This does not affect functionality
as those strings are intended for debugging only.

I also added .future tests in

  test/types/string/vass/string-in-tuple

where I stashed the unmodified copies of DimensionalDist2D
and BlockCycDim that can be exercised. Once the bug is fixed
and the original DimensionalDist2D and BlockCycDim could
be reverted to get strings back and this newly-added directory
can be removed altogether.