Don't count the waiting task in _waitEndCount as a running task
In _waitEndCount, we now  decrement the running task count before waiting for
all spawned tasks to complete. Then we increment after all tasks are done.

The original task will just be waiting and yielding so we alter the
runningTaskCount to better utilize all the cores.

This was discovered as part of the steam-ep performance work, where locale 0
(where the coforall occurred) had more running tasks. While it's technically
true that the task is still running it's just doing a waitFor(), which will
yield the processor so we might as well do some real work.

Now something like:

```chapel
coforall loc in Locales do on loc {
  writeln(loc.runningTasks());
}
```

should report 1 running task everywhere. It does always report 1 when using
network atomics. Unfortunately it's hard to make a portable test because the
remote tasks are going to execute an on statement back to locale 0 to do a
_downEndCount. That _downEndCount will mess with the running tasks on locale 0,
so we can't make a reliable test. For now, we'll rely on our performance tests
to lock in the behavior.