Remove specified return type check.
Declared return types are now supported for procs (though not entirely for
iters) as of PR #2322.

This check was added because it was more helpful to let --verify catch the
error than to debug the resulting memory management errors.  Now the check
inhibits compiling most of the String module with --verify enabled, so the
check is removed.