Small changes to get GASNet and dlmalloc building with CCE with -hgnu
CCE had trouble with a static int pointer to const in GASNet, so remove the
const qualifier when compiling with CCE.

CCE doesn't support the '__sync_lock_release' intrinsic that dlmalloc uses
for locking for gcc, so choose a different locking implementation.

Tested on release/examples with CHPL_COMM=gasnet, CHPL_MEM=dlmalloc and
PrgEnv-cray.