Load the target compiler for whitebox testing
The internal load_compiler function was split into load_gen_compiler and
load_target_compiler versions. This just updates references to it in the public
repo.

The module uses the gen compiler to build the compiler and runtime, and the
target version to test. For whitebox testing we use the target compiler for
everything because there's no easy way to split up what we build with vs test
with. We decided to always use the target compiler to get more exposure
building with newer compilers.