Disable _tracedd functions.
This is a workaround for the performance regression seen on HPL.  The purpose
of the workaround is to gauge the potential impact of the real solution.

In string-as-rec, these trace routines are not completely inlined out, even
though evaluation of the condition in the leaf routine means that it will have
no observable side-effects.  The performance degradation is seen because the
new string implementation is a record, so its autoCopy routine must be called
when it is copied into a tuple field.  In the current implementation of the
tracedd routines, this happens a lot, because they are written as varargs functions.

In the current implementation, varargs expansion (such as "...args") is
performed using a tuple.  It is conjectured that removing the tuple will
eliminate the autoCopy calls that are currently required in filling its
fields.  If this conjecture is true, then following the implementation change,
the _traceddc and _traceddd calls can be re-enabled.