Fix chapel environment macros in runtime
90acdd82b49a9a7aacc0038e1a684ecedf0a65c5 added functionality that #defines
macros to represent the environment that the runtime was built with. i.e. the
chplenv is made available to the runtime as macros.

It basically took `printchplenv --simple`, replaces the '=' with a '_' and
converted to uppercase. This resulted in things like CHPL_TARGET_COMPILER_GNU
being defined. However, we didn't realize that some chplenv values contain '-'s
and those aren't legal identifies for macros. This just replaces all '-'s with
'_'s so something like CHPL_TARGET_COMPILER_CRAY-PRGENV-GNU is now
CHPL_TARGET_COMPILER_CRAY_PRGENV_GNU.

In the future we might want to remove all non [A-Za-z0-9_] characters