Delete the string-as-rec branch before checking it out
We have some hacks in place to test the string-as-rec branch in a few
configurations. However, all of the repo management is done by jenkins and it
doesn't know that its being lied to about what branch to test so it's never
updating string-as-rec

It's hard to have jenkins use a different branch, so just delete the branch
before checking it out to make sure we get an up to date copy.