Add tolerances to the remaining tests in users/aroonsharma
Most of the tests in this directory do correctness tests comparing floating
point values. Only a few of them allowed for some tolerance using a
"within_epsilon" function. The rest tried to do direct floating point
comparison using '=='.

A few of these that do direct comparison started failing with a83daf117479.
That commit corrected the running task count and some of the reductions were
actually being done in parallel. Previously they were serial so the order the
floating point operations were done was the same. Since they are done in
parallel now, we have to add a tolerance.

This adds a tolerance to all the tests that did floating point comparison not
just the ones that were failing. I used a cleaned up copy of the
"within_epsilon" that a few of the tests already used.