Fix this future so that its behavior more accurately reflects our vision
And remove the .future and .bad files, since we believe the state of the
compiler is correct.  Basically, we expect string literals to be immutable,
meaning they can't be passed as out, inout, or ref arguments to functions.  And
since literals are no longer c_string, only the new last section exercises
c_strings to these functions.  We also believe that c_strings should not be
implicitly converted to strings when passed to functions, so the .bad error
output is now correct output.