Fix chpl-env-gen test for pgi
The test included (-I) the runtime gen path so it could find the chpl-env-gen.h
file. However, on the gen path for whitebox testing is really long and pgi
silently truncates so we weren't able to find the file.

Now we just copy the file to the test directory instead of including it.