Workaround for the crash in fileNotValid.
In the test modules/standard/Path/lydia/realpath/fileNotValid.chpl, there is a
crash because the autoDestroy for the temporary result of the call to string()
is being inserted in the wrong place.  There is a pivotal item that recommends
a better solution (#101646244).

In the mean time, the problem can be worked around by inserting an explicit
"else" clause into the source code.  This patch contains a new future test,
also to track the error and recommend how it is to be resolved.