Use testEnv to evaluate executable skipifs as well
Previously testEnv only processed our special skipif language. When we added
executable skipifs, we ran them directly instead of calling testEnv. This
occurred twice in start_test, once in paratest, and once in sub_test.

However, it was only in sub_test that we stuffed the chplenv in the environment
of the executable skipifs. This meant that directory level skipifs couldn't
see inferred chplenv values. We want to be able to see the values of inferred
chplenv variables in any skipif. I originally started adding the chplenv to the
executable skipif calls in all our scripts, but quickly realized how many
places we were doing that. Instead, it seems simpler and cleaner to just let
testEnv detect if the file is executable. It already stuffs the chplenv in the
env so there's no need to duplicate that logic elsewhere.

This commit has testEnv execute the file instead of running our skipif language
processing on it if the file is executable. Otherwise it's business as usual.
I them removed all the special checks for executable skipif files in
start_test, paratest, and sub_test since testEnv does the check itself now.