Resolve some passing futures on string-as-rec branch
functions/deitz/nested/test_nested_var_iterator3 failed due to copy propagation
not being able to follow references through record fields.  This now passes on
the string-as-rec branch.  Tom says:

"To support allowing iterators to update record types in general, I reduced the number of cases where the iterator record and class force reference arguments to be stored by value.  These are
  - variables marked with the FLAG_TEMP_IN_ITERATOR
  - variables whose base type is marked with FLAG_ITERATOR_RECORD
  - The _this member of an iterator method, unless the method's "this" intent includes "ref"."

variables/vass/ref-to-domains-arrays previously failed because when a ref to a
domain, array or slice was made, the reference count decrements were correct but
the increments weren't.  This now passes.  Tom says:

"Changes were made in several places to get the reference counts right."

I verified that some of the settings specified by the skipif also work, since it
was specified that their output was different than the base case.  Specifically,
I verified that no-local, valgrind, and comm=gasnet testing passed.  There are
other configurations skipped, but I think we'll be fine verifying those when
this change goes in.